pyproject.toml
[tool.poetry]
name = "aldentefier-api"
version = "0.1.0"
description = "Web API for the Aldentefier model."
authors = ["Dominic Thorn <dominic.thorn@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.61.1"
uvicorn = "^0.12.2"
torch = "^1.6.0"
pillow = "^8.0.1"
fastai = "^2.0.16"
IPython = "^7.18.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.4"
pytest = "^6.1.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
